requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
bitsandbytes==0.42.0
certifi==2025.10.5
chardet==3.0.4
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
discord.py==2.6.4
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
fastapi==0.119.0
ffmpy==0.6.3
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
googletrans==4.0.0-rc5
gradio==4.16.0
gradio_client==0.8.1
h11==0.9.0
h2==3.2.0
hf-xet==1.1.10
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.35.3
hyperframe==5.2.0
idna==2.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latex2mathml==3.78.1
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
narwhals==2.8.0
networkx==3.4.2
numpy==1.26.4
orjson==3.11.3
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==10.4.0
propcache==0.4.1
psutil==7.1.0
pydantic==2.12.2
pydantic_core==2.41.4
pydantic-settings==2.12.0
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rfc3986==1.5.0
rich==14.2.0
rpds-py==0.27.1
ruff==0.14.0
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
svgwrite==1.4.3
sympy==1.14.0
threadpoolctl==3.6.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.37.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
wavedrom==2.0.3.post3
websockets==11.0.3
yarl==1.22.0
aiosqlite>=0.20
python-multipart==0.0.20